requirements/base.txt
asgiref==3.11.1
Django==6.0.4
psycopg2-binary==2.9.10
sqlparse==0.5.5
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
factory-boy==3.3.3
django-filter==25.2